+++ pywbem_mock/_wbemconnection_mock.py
@@ -51,6 +51,9 @@ from pywbem import WBEMConnection, CIMClass, CIMClassName, \
 from pywbem._nocasedict import NocaseDict
 from ._dmtf_cim_schema import DMTFCIMSchema
 
+if six.PY2:
+    import codecs  # pylint: disable=wrong-import-order
+
 
 __all__ = ['FakedWBEMConnection', 'method_callback_interface']
 
@@ -97,13 +100,21 @@ def _uprint(dest, text):
     If dest is None, the text is encoded to a codepage suitable for the current
     stdout and is written to stdout.
     """
+    assert isinstance(text, six.text_type)
     if dest is None:
         btext = text.encode(STDOUT_ENCODING, 'replace')
-        print(btext)
+        if six.PY3:
+            print(btext.decode(STDOUT_ENCODING))
+        else:
+            print(btext)
     else:
-        btext = text.encode('utf-8')
-        with open(dest, 'a') as f:
-            print(btext, file=f)
+        if six.PY2:
+            # Open with codecs to define text mode
+            with codecs.open(dest, mode='a', encoding='utf-8') as f:
+                print(text, file=f)
+        else:
+            with open(dest, 'a', encoding='utf-8') as f:
+                print(text, file=f)
 
 
 def method_callback_interface(conn, objectname, methodname, **params):
@@ -760,7 +771,7 @@ class FakedWBEMConnection(WBEMConnection):
             raise ValueError('Invalid output format definition %s. '
                              '%s are valid.' % (output_format, OUTPUT_FORMATS))
 
-        _uprint(dest, '%s========Mock Repo Display fmt=%s namespaces=%s '
+        _uprint(dest, u'%s========Mock Repo Display fmt=%s namespaces=%s '
                       '=========%s\n' %
                 (cmt_begin, output_format,
                  ('all' if namespaces is None else namespaces), cmt_end))
@@ -785,7 +796,7 @@ class FakedWBEMConnection(WBEMConnection):
         repo_nss = sorted(repo_nss)
 
         for ns in repo_nss:
-            _uprint(dest, '\n%sNAMESPACE %s%s\n' % (cmt_begin, ns, cmt_end))
+            _uprint(dest, u'\n%sNAMESPACE %s%s\n' % (cmt_begin, ns, cmt_end))
             self._display_objects('Qualifier Declarations', self.qualifiers,
                                   ns, cmt_begin, cmt_end, dest=dest,
                                   summary=summary, output_format=output_format)
@@ -799,7 +810,7 @@ class FakedWBEMConnection(WBEMConnection):
                                   cmt_begin, cmt_end, dest=dest,
                                   summary=summary, output_format=output_format)
 
-        _uprint(dest, '============End Repository=================')
+        _uprint(dest, u'============End Repository=================')
 
     @staticmethod
     def _display_objects(obj_type, objects_repo, namespace, cmt_begin, cmt_end,
@@ -813,12 +824,12 @@ class FakedWBEMConnection(WBEMConnection):
         # TODO:ks FUTURE Consider sorting to perserve order of compile/add.
         if namespace in objects_repo:
             if obj_type == 'Methods':
-                _uprint(dest, '%sNamespace %s: contains %s %s:%s\n' %
+                _uprint(dest, u'%sNamespace %s: contains %s %s:%s\n' %
                         (cmt_begin, namespace,
                          len(objects_repo[namespace]),
                          obj_type, cmt_end))
             else:
-                _uprint(dest, '%sNamespace %s: contains %s %s %s\n' %
+                _uprint(dest, u'%sNamespace %s: contains %s %s %s\n' %
                         (cmt_begin, namespace,
                          len(objects_repo[namespace]), obj_type, cmt_end))
             if summary:
@@ -833,14 +844,14 @@ class FakedWBEMConnection(WBEMConnection):
                 # TODO:ks Future: Possibly sort insts by path order.
                 for inst in insts:
                     if output_format == 'xml':
-                        _uprint(dest, '%s Path=%s %s\n%s' %
+                        _uprint(dest, u'%s Path=%s %s\n%s' %
                                 (cmt_begin, inst.path.to_wbem_uri(), cmt_end,
                                  _pretty_xml(inst.tocimxmlstr())))
                     elif output_format == 'repr':
-                        _uprint(dest, 'Path:\n%r\nInst:\n%r\n' %
+                        _uprint(dest, u'Path:\n%r\nInst:\n%r\n' %
                                 (inst.path, inst))
                     else:
-                        _uprint(dest, '%s Path=%s %s\n%s' %
+                        _uprint(dest, u'%s Path=%s %s\n%s' %
                                 (cmt_begin, inst.path.to_wbem_uri(), cmt_end,
                                  inst.tomof()))
 
@@ -853,7 +864,7 @@ class FakedWBEMConnection(WBEMConnection):
 
                 for cln in methods:
                     for method in methods[cln]:
-                        _uprint(dest, '%sClass: %s, method: %s, '
+                        _uprint(dest, u'%sClass: %s, method: %s, '
                                       'callback: %s %s' %
                                 (cmt_begin, cln, method,
                                  methods[cln][method].__name__, cmt_end))
@@ -868,7 +879,7 @@ class FakedWBEMConnection(WBEMConnection):
                     if output_format == 'xml':
                         _uprint(dest, _pretty_xml(obj.tocimxmlstr()))
                     elif output_format == 'repr':
-                        _uprint(dest, '%r' % obj)
+                        _uprint(dest, u'%r' % obj)
                     else:
                         _uprint(dest, obj.tomof())